[backend/cmd/watch] Refactored to use a state machine. (#15)
The previous code of the watch command was very hard to follow as the state was all over the place. The current implementation allows you to reason about the whole chain a lot easier. It gives a clear overview of what can happen next and will give an error when an invalid state is reached. In case that state is still reached, it means something wasn't thought of and it can be added very easily.

Enabling verbose output also allows you to debug it relatively easily as it gives a good overview of the inputs and most of the times logs that come from the states themselves.