Bumbed mysql version
Using mysql version 8.X to work with the backend, setting legecy login
true for compliency. The migrations run after the db has started but not
after it has finsiht setting up this can be annoying the first time you
buids as the db takes quite some time to  build all the tables and the
migrations keeps trying to execute and fail until it can do all the
migrations.