Added OrderStatus Pickup
"Ophalen" as extra status for orders. Buyers that check their order (email link) will the see the status as 'ophalen' and know they can come to collect.
This extra status option is also needed/useful for webshops that do not keep stock. 'Prepared'(=verwerkt) as a status would then mean it is passed on to the supplier.