Update OrderView.vue
The hasWrite function already existed here, but it did not work. Read-only users could still accept payments.
I changed this one line of code to a snippet I found somewhere else. Without really undestanding how the original piece of code or the new piece exactly work. :)
The issue however may also lie in the fact that a user that does not have access to payments, somehow still has. And therefore incorrectly validates as true in this piece of code as 'canManagePayments'. In that case my replacement here is futile and the issue with the payments access rights needs to be resolved to also fix this.