Reworked permission checking via request contexts
no issue

This cleans up the permission checking code by removing it
from all endpoint code to a single class that is used to
check whether the current user has access to a resource.

Apart from that, a new context is introduced so there
is no need to pass the current organization and user around.

This will help with the move towards users that are
scoped to multiple organizations.