Fixed reusing of member manager when logging out and back in
fixes STA-91

Now we force a replace to the auto login root if you are no longer signed in. Next time you sign in, the auto root will show, which will be determined automatically based on the user. But that will also cause the getRootView methods to be reevaluated, and a new member manager will be created.