Fixed password forgot view not working in apps
fixes STA-221

This was caused by the update view that was displayed shortly before the password reset view, and then this view dismissed the password forgot view instead of itself, resulting in an infinite spinner that was visible.

This change also prevents further race conditions and prevents a duplicate update check on launching the app.