Updated Alamofire to 5.10.2
Capacitor doesn't update Pods by default, so you'll need to run:
- `pod outdated` manually to list all updateable dependencies
- run `pod update Alamofire` for all outdated dependencies that can be updated to a newer version.

This also updates `@capgo/capacitor-updater` to the latest version for Capacitor v6.

There was an issue in the capacitor update dependency that made the upgrade a bit harder because it didn't reference the right minimum version for Alamofire in its podfile.