Removed reactivity from user permissions
This fixes possible infinite rendering loops because calculating the permissions sometimes alters it dependencies. I couldn't fix that for some reason, so I settled on never using the proxies inside the calculations and keeping track of the required caches to reduce the amount of permission calculations to the very minimum.

This should have a positive impact on the performance of the application.