Updated translation tool to not sync translations and always use the same settings
The env should only define the default locale, but the other settings should always stay the same and should be committed to git.

We don't want to add Dutch translations for example to the en.json files, because that polutes it and makes it harder to find which translations are already translated and which are. E.g. all translations in en.json are already translated, but suddenly we add 10 new Dutch translations -> how do you know which ones are already translated programmatically?