Fixed ignoring locked memberships while creating new memberships
fixes STA-707

A member could have a locked membership for a different organization that it should have got a membership for. E.g. a member was registered at organization A, then charged, and then moved to organization B.

If that happens, we should prevent creating a new membership for organization B because we cannot delete the membership for organization A. In case the membership should move, an admin should unlock the old membership and it will automatically move to the new organization - creating the required balance changes if it has been charged.