Fixed permission edge case
Context:
A platform admin only has read permissions to a certain tag (= makes sure this platform admin can only access the current period, not previous ones)

Problem:
The admin loads the platform members table, but it throws an error in this specific edge case.

Cause:
This table contains members that have registration with a period id that doesn't match the period id of the corresponding group. This happened because the group's period id was changed after a registration happened (when the date of an event was changed). When we load members, we filter by registration.periodId, but when we check the member permissions, we check group.periodId

This causes an issue that you cannot load the members table any more because it contains a member where canAccessMember returns false (the scope filter doesn't match the canAccessMember method).

Fix:
Filter the scope by group.periodId instead of registration.periodId