Fixed allow platform admins to access members from the current platform period if the organization period is not the same
When checking permissions to a specific member, we always use the organization periodId, but when we filter the scope for platform admins, we use the platform's period id. To fix this difference, we also check permission to a member (canAccessMember calls canAccessGroup indirectly) by allowing access to a group if the group has the same period id as the platform.