Added queues for sessions to prevent refreshing same token multiple times
fixes STA-721

Multiple refreshes could happen on page load. This caused an issue when there were too many refresh tokens for a user - where the backend would delete the tokens that will expire the earliest. Because of the duplicate calls, one of those calls could end up with a deleted token, logging out the user in a race condition