Limited record answer access to currently scoped organization
When the backend checks whether you can access a record or not, it now only checks the currently scoped organization. For platform admins, we make an exception and also check other organizations (in case the admin has the record permission for at least one organization where the member is registered).

In the past, the backend checked all the organizations the user had permission for - but this isn't correct since the user should also have permission to read or edit that member for the organization that granted permission to a specific record.