Fixed possible race condition when storing fetched orders
fixes STA-825

The receive handlers were called in parallel. That could cause issues if executed in the wrong order. On top of that, the storing of the timestamps happened in parallel with storing the orders, but should happen after storing the orders to be sure that everything is stored correctly without errors.