Fixed scroll position forgotten when presenting modally
fixes STA-846

Cause:
When presenting modally, a different navigation controller, higher up the dom tree is used. This navigation controller searches for the upper level <main> element to remember the scroll position, but this is the tab bar controller's main element instead of the main element of an st-view.

Theoretically, we should start supporting remembering the scroll position of multiple views in the navigation controller's implementation.

For now, that isn't a requirement because we never present anything on top of a view with multiple scroll views